requirements.txt
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0
datasets>=2.21.0
evaluate>=0.4.0
wandb>=0.17.0
tensorboard>=2.17.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pyyaml>=6.0.1
scikit-learn>=1.5.0